+++ src/NeatMapper/Extensions/MatcherExtensions.cs
@@ -422,7 +422,345 @@ namespace NeatMapper {
 		#endregion
 
 		#region Predicate
-		
+		#region Runtime
+		#region Source
+		/// <summary>
+		/// Creates a factory which can be used as a predicate to match multiple objects against a single one.
+		/// </summary>
+		/// <param name="source">Object to compare other objects to, may be null.</param>
+		/// <param name="sourceType">Type of the object to compare to, used to retrieve the available maps.</param>
+		/// <param name="destinationType">
+		/// Type of the other objects to compare with <paramref name="source"/>, used to retrieve the available maps.
+		/// </param>
+		/// <param name="mappingOptions">
+		/// Additional options passed to the context, support depends on the matcher and/or the maps, null to ignore.
+		/// </param>
+		/// <returns>
+		/// A factory which can be used as a predicate to compare objects of type <paramref name="destinationType"/>
+		/// with the provided object <paramref name="source"/>.
+		/// </returns>
+		/// <exception cref="MapNotFoundException">The provided types could not be matched.</exception>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			source,
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (matcher == null)
+				throw new ArgumentNullException(nameof(matcher));
+			if (sourceType == null)
+				throw new ArgumentNullException(nameof(sourceType));
+			if (destinationType == null)
+				throw new ArgumentNullException(nameof(destinationType));
+
+			var factory = matcher.MatchFactory(sourceType, destinationType, mappingOptions);
+			return new DisposableNewMapFactory<object, bool>(destination => factory.Invoke(source, destination), factory);
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <inheritdoc cref="Predicate(IMatcher, object, Type, Type, MappingOptions)"/>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			source,
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			IEnumerable?
+#else
+			IEnumerable
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate(source, sourceType, destinationType, mappingOptions != null ? new MappingOptions(mappingOptions) : null);
+		}
+
+		/// <inheritdoc cref="Predicate(IMatcher, object, Type, Type, MappingOptions)"/>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			source,
+			Type sourceType,
+			Type destinationType,
+			params
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?[]?
+#else
+			object[]
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate(source, sourceType, destinationType, mappingOptions?.Length > 0 ? new MappingOptions(mappingOptions) : null);
+		}
+		#endregion
+
+		#region Destination
+		/// <summary>
+		/// Creates a factory which can be used as a predicate to match multiple objects against a single one.
+		/// </summary>
+		/// <param name="sourceType">
+		/// Type of the other objects to compare with <paramref name="destination"/>, used to retrieve the available maps.
+		/// </param>
+		/// <param name="destination">Object to compare other objects to, may be null.</param>
+		/// <param name="destinationType">Type of the object to compare to, used to retrieve the available maps.</param>
+		/// <param name="mappingOptions">
+		/// Additional options passed to the context, support depends on the matcher and/or the maps, null to ignore.
+		/// </param>
+		/// <returns>
+		/// A factory which can be used as a predicate to compare objects of type <paramref name="sourceType"/>
+		/// with the provided object <paramref name="destination"/>.
+		/// </returns>
+		/// <exception cref="MapNotFoundException">The provided types could not be matched.</exception>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+			Type sourceType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			destination,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (matcher == null)
+				throw new ArgumentNullException(nameof(matcher));
+			if (sourceType == null)
+				throw new ArgumentNullException(nameof(sourceType));
+			if (destinationType == null)
+				throw new ArgumentNullException(nameof(destinationType));
+
+			var factory = matcher.MatchFactory(sourceType, destinationType, mappingOptions);
+			return new DisposableNewMapFactory<object, bool>(source => factory.Invoke(source, destination), factory);
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <inheritdoc cref="Predicate(IMatcher, Type, object, Type, MappingOptions)"/>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+			Type sourceType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			destination,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			IEnumerable?
+#else
+			IEnumerable
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate(sourceType, destination, destinationType, mappingOptions != null ? new MappingOptions(mappingOptions) : null);
+		}
+
+		/// <inheritdoc cref="Predicate(IMatcher, Type, object, Type, MappingOptions)"/>
+		public static NewMapFactory<object, bool> Predicate(this IMatcher matcher,
+			Type sourceType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			destination,
+			Type destinationType,
+			params
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?[]?
+#else
+			object[]
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate(sourceType, destination, destinationType, mappingOptions?.Length > 0 ? new MappingOptions(mappingOptions) : null);
+		}
+		#endregion
+		#endregion
+
+		#region Explicit destination, inferred source
+		/// <summary>
+		/// Creates a factory which can be used as a predicate to match multiple objects against a single one.
+		/// </summary>
+		/// <param name="source">
+		/// Object to compare other objects to, CANNOT be null as the source type will be retrieved from it at runtime,
+		/// which will be used to retrieve the available maps
+		/// </param>
+		/// <param name="mappingOptions">
+		/// Additional options passed to the context, support depends on the matcher and/or the maps, null to ignore.
+		/// </param>
+		/// <returns>
+		/// A factory which can be used as a predicate to compare objects of type <typeparamref name="TDestination"/>
+		/// with the provided object <paramref name="source"/>.
+		/// </returns>
+		/// <exception cref="MapNotFoundException">The provided types could not be matched.</exception>
+		public static NewMapFactory<TDestination, bool> Predicate<TDestination>(this IMatcher matcher,
+			object source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (matcher == null)
+				throw new ArgumentNullException(nameof(matcher));
+			if (source == null)
+				throw new ArgumentNullException(nameof(source), "Type cannot be inferred from null source, use an overload with an explicit source type");
+
+			var factory = matcher.MatchFactory(source.GetType(), typeof(TDestination), mappingOptions);
+			return new DisposableNewMapFactory<TDestination, bool>(destination => factory.Invoke(source, destination), factory);
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <inheritdoc cref="Predicate{TDestination}(IMatcher, object, MappingOptions)"/>
+		public static NewMapFactory<TDestination, bool> Predicate<TDestination>(this IMatcher matcher,
+			object source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			IEnumerable?
+#else
+			IEnumerable
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate<TDestination>(source, mappingOptions != null ? new MappingOptions(mappingOptions) : null);
+		}
+
+		/// <inheritdoc cref="Predicate{TDestination}(IMatcher, object, MappingOptions)"/>
+		public static NewMapFactory<TDestination, bool> Predicate<TDestination>(this IMatcher matcher,
+			object source,
+			params
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?[]?
+#else
+			object[]
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate<TDestination>(source, mappingOptions?.Length > 0 ? new MappingOptions(mappingOptions) : null);
+		}
+		#endregion
+
+		#region Explicit source and destination
+		/// <summary>
+		/// Creates a factory which can be used as a predicate to match multiple objects against a single one.
+		/// </summary>
+		/// <param name="source">Object to compare other objects to, may be null.</param>
+		/// <param name="mappingOptions">
+		/// Additional options passed to the context, support depends on the matcher and/or the maps, null to ignore.
+		/// </param>
+		/// <returns>
+		/// A factory which can be used as a predicate to compare objects of type <typeparamref name="TDestination"/>
+		/// with the provided object <paramref name="source"/>.
+		/// </returns>
+		/// <exception cref="MapNotFoundException">The provided types could not be matched.</exception>
+		public static NewMapFactory<TDestination, bool> Predicate<TSource, TDestination>(this IMatcher matcher,
+#if NET5_0_OR_GREATER
+			TSource?
+#else
+			TSource
+#endif
+			source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable disable
+#endif
+
+			if (matcher == null)
+				throw new ArgumentNullException(nameof(matcher));
+
+			var factory = matcher.MatchFactory<TSource, TDestination>(mappingOptions);
+			return new DisposableNewMapFactory<TDestination, bool>(destination => factory.Invoke(source, destination), factory);
+
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+#nullable enable
+#endif
+		}
+
+		/// <inheritdoc cref="Predicate{TDestination}(IMatcher, object, MappingOptions)"/>
+		public static NewMapFactory<TDestination, bool> Predicate<TSource, TDestination>(this IMatcher matcher,
+#if NET5_0_OR_GREATER
+			TSource?
+#else
+			TSource
+#endif
+			source,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			IEnumerable?
+#else
+			IEnumerable
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate<TSource, TDestination>(source, mappingOptions != null ? new MappingOptions(mappingOptions) : null);
+		}
+
+		/// <inheritdoc cref="Predicate{TDestination}(IMatcher, object, MappingOptions)"/>
+		public static NewMapFactory<TDestination, bool> Predicate<TSource, TDestination>(this IMatcher matcher,
+#if NET5_0_OR_GREATER
+			TSource?
+#else
+			TSource
+#endif
+			source,
+			params
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?[]?
+#else
+			object[]
+#endif
+			mappingOptions) {
+
+			return matcher.Predicate<TSource, TDestination>(source, mappingOptions?.Length > 0 ? new MappingOptions(mappingOptions) : null);
+		}
+		#endregion
 		#endregion
 	}
 }