+++ src/NeatMapper/Mappers/Core/MergeCollectionMapper.cs
@@ -334,32 +334,26 @@ namespace NeatMapper {
 											// Deleted elements
 											if (removeNotMatchedDestinationElements) {
 												foreach (var destinationElement in destinationEnumerable) {
-													bool found = false;
-													foreach (var sourceElement in sourceEnumerable) {
-														if (elementsMatcherFactory.Invoke(sourceElement, destinationElement)) {
-															found = true;
-															break;
-														}
-													}
+													if (!sourceEnumerable.Cast<object>()
+																.Any(sourceElement => elementsMatcherFactory.Invoke(sourceElement, destinationElement))) {
 
-													if (!found)
 														elementsToRemove.Add(destinationElement);
+													}
 												}
 											}
 
 											// Added/updated elements
 											foreach (var sourceElement in sourceEnumerable) {
+												// Cannot use FirstOrDefault because there might be matching null elements
 												bool found = false;
 												object matchingDestinationElement = null;
-												foreach (var destinationElement in destinationEnumerable) {
-													if (elementsMatcherFactory.Invoke(sourceElement, destinationElement) &&
-														!elementsToRemove.Contains(destinationElement)) {
-
-														matchingDestinationElement = destinationElement;
-														found = true;
-														break;
-													}
+												try {
+													matchingDestinationElement = destinationEnumerable.Cast<object>()
+														.First(destinationElement => elementsMatcherFactory.Invoke(sourceElement, destinationElement) &&
+															!elementsToRemove.Contains(destinationElement));
+													found = true;
 												}
+												catch {}
 
 												if (found) {
 													// Try merge map